+++ R/overlap_percent.R
@@ -80,7 +80,7 @@ overlap_percent <- function(peaklist1,
     # When peaklist1 abd peaklist2 have multiple elements
   } else { 
       messager("Computing precision-recall results.")
-      #### peaklist1 as the query, peaklist2 as the subject ####
+      #### peaklist2 as the query, peaklist1 as the subject ####
       percent_list1 <- mapply(peaklist1, 
                              SIMPLIFY = FALSE,
                              FUN = function(x){
@@ -88,17 +88,17 @@ overlap_percent <- function(peaklist1,
                  SIMPLIFY = FALSE,
                  FUN=function(y){
                      overlap <- f(
-                         IRanges::subsetByOverlaps(x = x,
-                                                   ranges = y,
+                         IRanges::subsetByOverlaps(x = x, # subject
+                                                   ranges = y, # query
                                                    invert = invert)
                      )
                      percent <- length(overlap)/length(x)*100
                      data.table::data.table(Percentage=signif(percent,3)) 
          }) |> data.table::rbindlist(use.names = TRUE, idcol = "subject")
       }) |> data.table::rbindlist(use.names = TRUE, idcol = "query")
-      percent_list1$type <- "precision"
+      percent_list1$type <- "recall"
       
-      #### peaklist2 as the query, peaklist1 as the subject ####
+      #### peaklist1 as the query, peaklist2 as the subject ####
       percent_list2 <- mapply(peaklist2, 
                               SIMPLIFY = FALSE,
                               FUN = function(x){
@@ -106,15 +106,15 @@ overlap_percent <- function(peaklist1,
                  SIMPLIFY = FALSE,
                  FUN=function(y){
                      overlap <- f(
-                         IRanges::subsetByOverlaps(x = x,
-                                                   ranges = y,
+                         IRanges::subsetByOverlaps(x = x, # subject 
+                                                   ranges = y, # query
                                                    invert = invert)
                      )
                      percent <- length(overlap)/length(x)*100
                      data.table::data.table(Percentage=signif(percent,3)) 
                  }) |> data.table::rbindlist(use.names = TRUE, idcol = "subject")
     }) |> data.table::rbindlist(use.names = TRUE, idcol = "query")
-      percent_list2$type <- "recall"
+      percent_list2$type <- "precision"
       df <- rbind(percent_list1, percent_list2)
       id <- type <- query <- subject <- NULL;
       df[,peaklist1:=ifelse(type=="precision",query,subject)]